coder-finetune Tier 8.1: bug fixes (extract_code prose, async-def format reward, eval pass@k + eos, dtype kwarg, GRPO divisibility) (24 → 37 tests)
extract_code: prose-prefixed completions ("Sure, here you go:\ndef ...")
returned the whole string with prose, then exec() died with SyntaxError and
the verifier scored 0. Now we recover the code suffix from the first
def/async def/class/import line.

format reward _DEF_RE: didn't match `async def`, so async helpers lost the
format bonus even with a clean fenced block. Added async-def to the regex.

eval/run_humaneval.py: --n-samples was a no-op (argparse accepted it,
quick_eval ignored it) — pass@k silently ran as pass@1. Now uses
num_return_sequences and short-circuits on the first pass. Also added the
missing eos_token_id (eval ran past <|im_end|> to max_new_tokens) and a
--seed flag for reproducible eval.

infer/generate.py: dtype= breaks on transformers 4.45 (in our pin range).
Switched to torch_dtype= for full-range compat (matches merge_lora.py +
train.py).

GRPO: shipped grpo_3050.yaml had num_generations=6 against an effective
batch of 8 (bs=2 * accum=4 mod 6 = 2) — TRL would later raise an opaque
shape error deep in training. Added a divisibility validator that fails
fast with an actionable message, and fixed the shipped config to G=8.

+++ coder-finetune/infer/generate.py
@@ -44,6 +44,11 @@ def _load(model_path: str, device: str, dtype):
     from transformers import AutoModelForCausalLM, AutoTokenizer
     p = Path(model_path)
     is_adapter = (p / "adapter_config.json").exists()
+    # transformers' load kwarg name changed: pre-4.46 only accepts ``torch_dtype``;
+    # 4.46+/5.x accept both (``dtype`` is the new spelling, ``torch_dtype`` still
+    # works as an alias). Our requirements.txt allows `>=4.45,<5.0`, so the older
+    # spelling is the safe one for the full pin range — merge_lora.py + train.py
+    # already use it. (The previous ``dtype=`` form would TypeError on 4.45.)
     if is_adapter:
         cfg = json.loads((p / "adapter_config.json").read_text())
         base_name = cfg.get("base_model_name_or_path")
@@ -51,11 +56,11 @@ def _load(model_path: str, device: str, dtype):
             raise ValueError(f"adapter at {p} has no base_model_name_or_path")
         from peft import PeftModel
         tok = AutoTokenizer.from_pretrained(model_path)
-        base = AutoModelForCausalLM.from_pretrained(base_name, dtype=dtype)
+        base = AutoModelForCausalLM.from_pretrained(base_name, torch_dtype=dtype)
         model = PeftModel.from_pretrained(base, model_path)
     else:
         tok = AutoTokenizer.from_pretrained(model_path)
-        model = AutoModelForCausalLM.from_pretrained(model_path, dtype=dtype)
+        model = AutoModelForCausalLM.from_pretrained(model_path, torch_dtype=dtype)
     if tok.pad_token is None:
         tok.pad_token = tok.eos_token
     return tok, model.to(device).eval()
